requirements.txt
PyYAML==5.4.1
web3==5.20.0
